move to MiscCommands.py, rip say.py
